# server/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/houses": {
      "get": {
        "tags": [
          "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/House"
            }
          }
        }
      },
      "post": {
        "tags": [
          "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/problem+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HouseDetail"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HouseDetail"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HttpValidationProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/problem+json",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HouseDetail"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HttpValidationProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HouseDetail"
            }
          }
        }
      }
    },
    "/houses/{houseId}": {
      "get": {
        "tags": [
          "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "produces": [
          "application/problem+json",
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "houseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HouseDetail"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "produces": [
          "application/problem+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "houseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/houses/{houseId}/bids": {
      "get": {
        "tags": [
          "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "produces": [
          "application/problem+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "houseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/problem+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "houseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Bid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Bid"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HttpValidationProblemDetails"
            }
          },
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Bid": {
      "required": [
        "bidder"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "houseId": {
          "format": "int32",
          "type": "integer"
        },
        "bidder": {
          "type": "string"
        },
        "amount": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "House": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "HouseDetail": {
      "required": [
        "address",
        "country"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "address": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "photo": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "HttpValidationProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "errors": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "additionalProperties": { }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    }
  }
}